python/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
pymupdf
sentence-transformers
chromadb
python-dotenv
httpx
python-multipart